requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.1
Django==4.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-mathfilters==1.0.0
Faker==13.12.0
Pillow==9.2.0
psycopg2 @ file:///C:/ci_310/psycopg2_1642085998310/work
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
tzdata==2022.2
wincertstore==0.2
